Allow individual checks to be disabled by user, per #1091
Creates list of individual subchecks which have been disabled by user in config
file and then removes them from the list of checks to be executed. Each
subcheck's function should start with 'check_'. The name of the check should
also match the function. For example: The subcheck
typography.symbols.curly_quotes would be the function check_curly_quotes in
typography.symbols. I saw that some subchecks have slightly different function
names and so these would need to be adjusted. An example is the subcheck
typography.symbols.copyright, which is the function check_copyright_symbol in
typography.symbols. To keep the tests running, I think this subcheck's name
should be adjusted to typography.symbols.copyright_symbol.