refactor: raise minimum version from python 3.8 -> python 3.9
This commit is expected to land at a time close to or after 2024-10-01,
following which Python 3.8 will be end-of-life. Several tools already
require a Python version >=3.9, which has caused dependency resolution
issues for some time now. Python 3.9 changed the workings of the type
system  quite substantially. This will allow for a higher code quality
going forwards.

BREAKING CHANGE: Python 3.8 is no longer supported as of this commit.

To migrate, update your system's Python version.