refactor: remove unused unicode and dotall options
Applications of the unicode flag for checks were previously consistent,
and having it default to true makes no difference for any check within
proselint. Further, this behaviour is consistent with the default for
the Rust regex crate.

The only check that previously had dotall enabled did not use dot
matching, so it was redundant.

Removing these unnecessary features results in a cleaner codebase.