perf!: use dual engine system
Using google-re2 in place of python's re for as many checks as possible
resulted in a speedup of 33% on my system, alongside a 75% reduction in
the extra time taken to lint additional files.

It would be even faster to use pcre2 in place of python's re for checks
requiring backtracking and lookarounds, but the difference is minimal
currently, and pcre2 has no type stubs.

BREAKING CHANGE: This removes the `ignore_case` attribute in `Check`,
replacing it with a `Matcher` object that contains `RegexOptions`
alongside the engine setting.